+++ casia/src/components/organisms/Form/Form.js
@@ -4,6 +4,8 @@ import NoteForm from '../../molecules/NoteForm/NoteForm';
 const Form = props => {
   if (props.data === 'Note') {
     return <NoteForm />;
+  } else {
+    return null;
   }
 };