+++ casia/src/constants/index.js
@@ -1,0 +1,3 @@
+export default {
+  primaryYellow: '#ffea00'
+};