+++ casia/src/reducers/reducer.js
@@ -1,7 +1,8 @@
 const initalState = {
   counter: 1,
   creationItem: 'Note',
-  list: [{ id: 0, name: 'Matt' }]
+  list: [{ id: 0, name: 'Matt' }],
+  notes: [{ title: 'first note', context: 'Hello world' }]
 };
 
 const rootReducer = (state = initalState, action) => {
@@ -11,6 +12,13 @@ const rootReducer = (state = initalState, action) => {
       return {
         ...state,
         creationItem: (state.creationItem = action.payload.item)
+      };
+    }
+    case 'ADD_NOTE': {
+      console.log(action.payload.title);
+      return {
+        ...state,
+        notes: [...state.notes, action.payload]
       };
     }
     default: {