+++ casia/src/views/MenuViews/NotesView.js
@@ -1,13 +1,36 @@
 import React from 'react';
+import { connect } from 'react-redux';
+import styled from 'styled-components';
 
 import MainTemplate from '../../components/templates/MainTemplate/MainTemplate';
+import Note from '../../components/molecules/Note/note';
+
+const StyledGridWrapper = styled.div`
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  grid-gap: 10px;
+  width: 95%;
+  min-height: 60vh;
+`;
 
 const NotesView = props => {
   return (
     <MainTemplate>
-      <h1>This is notes View</h1>
+      {props.notes.length === 0 ? (
+        <h1>
+          You don't have any notes <span>💔</span>
+        </h1>
+      ) : (
+        <StyledGridWrapper>
+          {props.notes.map((val, i) => (
+            <Note title={val.title} context={val.context} />
+          ))}
+        </StyledGridWrapper>
+      )}
     </MainTemplate>
   );
 };
 
-export default NotesView;
+const mapStateToProps = ({ notes }) => ({ notes });
+
+export default connect(mapStateToProps)(NotesView);